audience-affinity-ondemand-flask-app/requirements.txt
# Install with: pip install -r requirements.txt

flask
pandas
faker
sentence-transformers
faiss-cpu
google-generativeai